+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2.3.0
+
+* Stretchable column
+  * `EasyTableModel`
+    * The `columnsWeight` method has been removed.
+  * `EasyTableColumn`
+    * The `weight` attribute has been renamed to `grow` and became nullable.
+  * `EasyTable`
+    * The `columnsFit` has been renamed to `columnWidthBehavior` and refactor from `bool` to `ColumnWidthBehavior` enum.
+
 ## 2.2.0
 
 * Bugfix